requirements.txt
azure-cognitiveservices-vision-customvision==1.0.0
